Big update: restructured ocean.js to support synchronous simulations.
All of the simulations are now held under the games array, and the
timer() function checks them all to see if there's anything to do in
them.

Remaining to do, but not essential, is to remove games from the array
once the simulation has ended and we've logged all its relevant
information.

Closes #23.