Working towards showing the prep text at the beginning.
Alright so the idea is that users will be shown a preparation text
while they're waiting for others to connect. Importantly, we should
ensure we're giving them enough time to read the text: they should
notify us when they have finished, by clicking a button, and once they
do, the system should wait a few seconds before beginning the task.

Currently, I've made it so the preparatory text appears when the
gameSettings.status value is 'waiting'. But what I really need to do is
to add another point to the workflow, between 'waiting' and
'running'---say 'getting ready'---that runs for a few seconds for all
players, still displaying the preparatory text as well as a note
telling them to get ready, because the simulation is about to begin.