Simulations now support many oceans!
A fairly big update, on which I still feel a bit wobbly but it seems to
be working OK so far. Simulations support multiple oceans, as specified
by the admin in newgroup.

One problem: we need a way to get back to the user to let them know
that there's no more space in the simulation, in case there is in fact
no space.

Closes #50.