Fix fisher index check
Simulations with all-humans would have the first human on spot 0 in the fishers array.
However, a number of methods (including readRules()) try to find the fisher in the array,
and for that first human, the index is 0. These methods then erroneously checked whether the
index was truthy, instead of checking whether it was not null. The consequence was that
the first fisher was never declared to be "ready", and even if they were, they would not
be able to interact with the simulation.

Closes #180.