Avoid triggering readRules() when starting tutorial
It looks like the behaviour for startng the tutorial was:
1. State that we've read the rules
2. Request pause
3. Show tutorial
4. Resume simulation

Presumably this was done to show all the elements of the page in the tutorial properly
(table of participants, ocean window, etc.). However, the pause/resume wasn't taking,
and the result was that the tutorial would carry on while the simulation was running.

This commit simplifies things by not stating that we've read the rules until after the
tutorial is done. The downside is that not all elements are visible when going through
the tutorial. I'll create a separate issue for that.